Рефакторит AppExceptionFilter
