--- src/shared/libs/rest/exception-filter/app-exception-filter.ts
@@ -1,38 +1,0 @@
-import { inject, injectable } from 'inversify';
-import { ExceptionFilter } from './exception-filter.interface.js';
-import { Component } from '../../../types/index.js';
-import { Logger } from '../../logger/index.js';
-import { Request, Response, NextFunction } from 'express';
-import { StatusCodes } from 'http-status-codes';
-import { HttpError } from '../index.js';
-import { createErrorObject } from '../../../helpers/common.js';
-
-@injectable()
-export class AppExceptionFilter implements ExceptionFilter {
-  constructor(@inject(Component.Logger) private readonly logger: Logger) {
-    this.logger.info('Register AppExceptionFilter');
-  }
-
-  private handleHttpError(error: HttpError, _req: Request, res: Response, _next: NextFunction) {
-    this.logger.error(`[${error.detail}]: ${error.statusCode} - ${error.message}`, error);
-    res
-      .status(error.statusCode)
-      .json(createErrorObject(error.message));
-  }
-
-  private handleOtherError(error: Error, _req: Request, res: Response, _next: NextFunction) {
-    this.logger.error(error.message, error);
-    res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .json({ message: error.message });
-  }
-
-  public catch(error: Error | HttpError, req: Request, res: Response, next: NextFunction): void {
-    if (error instanceof HttpError) {
-      this.handleHttpError(error, req, res, next);
-      return;
-    }
-
-    this.handleOtherError(error, req, res, next);
-  }
-}
