+++ src/shared/libs/rest/exception-filter/index.ts
@@ -1,3 +1,4 @@
 export * from './exception-filter.interface.js';
-export * from './app-exception-filter.js';
+export * from './app.exception-filter.js';
+export * from './validation.exception-filter.js';
 export * from './http-error.exception-filter.js';
